+++ lerna.json
@@ -12,6 +12,7 @@
     "packages/azpipelines/BuildTasks/DeployDestructiveManifestToOrgTask",
     "packages/azpipelines/BuildTasks/DeploySourceToOrgTask",
     "packages/azpipelines/BuildTasks/ExportSourceFromAnOrgTask",
+    "packages/azpipelines/BuildTasks/GenerateChangelogTask",
     "packages/azpipelines/BuildTasks/IncrementProjectBuildNumberTask",
     "packages/azpipelines/BuildTasks/InstallPackageDependenciesTask",
     "packages/azpipelines/BuildTasks/InstallSFDXCLITaskCurrent",
