+++ packages/azpipelines/BuildTasks/GenerateChangelogTask/task.json
@@ -1,0 +1,86 @@
+{
+    "id": "87d5162d-07ee-40ea-8bf4-ac8dfdb907f1",
+    "name": "sfpowerscripts-generatechangelog-task",
+    "friendlyName": "Generate Changelog",
+    "description": "Generates changelog for all the artifacts that is in a release, This task is part of  sfowerscripts",
+    "helpMarkDown": "",
+    "category": "Build",
+    "author": "dxatscale@accenture.com",
+    "version": {
+        "Major": 1,
+        "Minor": 0,
+        "Patch": 0
+    },
+    "runsOn": [
+        "Agent"
+    ],
+    "instanceNameFormat": "Generate a changelog for the current release",
+    "inputs": [
+        {
+            "name": "versionControlProvider",
+            "type": "pickList",
+            "label": "Select the version control provider",
+            "defaultValue": "github",
+            "options": {
+                "github": "GitHub",
+                "githubEnterprise": "GitHub Enterprise",
+                "bitbucket": "BitBucket Cloud",
+                "azureRepo": "Azure Repo",
+                "otherGit": "Other Git",
+                "hostedAgentGit": "Git which is already authenticated at the agent level"
+            },
+            "required": false,
+            "helpMarkDown": "Select a version control provider from the dropdown"
+        },
+        {
+            "name": "github_connection",
+            "type": "connectedService:github:OAuth,PersonalAccessToken,OAuth2",
+            "label": "GitHub Connection",
+            "defaultValue": "",
+            "required": false,
+            "helpMarkDown": "Specify the service connection name for your GitHub connection. Please note this service connection have permissions to read to the correspinding repostiory. Learn more about service connections [here](https://aka.ms/AA3am5s).",
+            "visibleRule": "versionControlProvider = github"
+        },
+        {
+            "name": "bitbucket_connection",
+            "type": "connectedService:bitbucket:OAuth,PersonalAccessToken,OAuth2",
+            "label": "BitBucket Connection",
+            "defaultValue": "",
+            "required": false,
+            "helpMarkDown": "Specify the service connection name for your BitBucket connection. Please note this service connection have permissions to read to the correspinding repostiory. Learn more about service connections [here](https://aka.ms/AA3am5s).",
+            "visibleRule": "versionControlProvider = bitbucket"
+        },
+        {
+            "name": "github_enterprise_connection",
+            "type": "connectedService:githubenterprise:OAuth,PersonalAccessToken,OAuth2",
+            "label": "GitHub Enterprise Connection",
+            "defaultValue": "",
+            "required": false,
+            "helpMarkDown": "Specify the service connection name for your Git Hub Enterprise connection. Please note this service connection have permissions to read from the correspinding repostiory. Learn more about service connections [here](https://aka.ms/AA3am5s).",
+            "visibleRule": "versionControlProvider = githubEnterprise"
+        },
+        {
+            "name": "username",
+            "type": "string",
+            "label": "Username",
+            "defaultValue": "",
+            "helpMarkDown": "Username for the Version Control Provider",
+            "required": true,
+            "visibleRule": "versionControlProvider = otherGit"
+        },
+        {
+            "name": "password",
+            "type": "string",
+            "label": "Password",
+            "defaultValue": "",
+            "required": true,
+            "helpMarkDown": "App Password/PAT for the Version Control Provider",
+            "visibleRule": "versionControlProvider = otherGit"
+        }
+    ],
+    "execution": {
+        "Node": {
+            "target": "lib/InstallPackageDependencies.js"
+        }
+    }
+}
