+++ packages/azpipelines/BuildTasks/GenerateChangelogTask/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../../tsconfig.json",
+   "compilerOptions": {
+        "outDir": "./lib", 
+        "sourceRoot": "./",
+        "composite": false
+    },
+    "files": [
+        "GenerateChangelog.ts"
+    ]
+  }
