+++ packages/core/src/PackageMetadata.ts
