+++ packages/core/src/generators/MDAPIPackageGenerator.ts
