+++ packages/core/src/manifest/ManifestHelpers.ts
