+++ packages/core/src/utils/OnExit.ts
